Remove all remaining devserver 2 support
